+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
 
 - Added Dart Extension Support 
+
+## 0.0.2
+
+- Updated Import settings
+
+## 0.0.3
+
+- Resolved Padding Extension conflicts